Fix extensions
