+++ 90s/90s/Extensions/Date+.swift
@@ -1,0 +1,16 @@
+//
+//  Date+.swift
+//  90s
+//
+//  Created by woong on 2021/03/27.
+//
+
+import Foundation
+
+extension Date {
+    func dateToString() -> String {
+        let dateFormatter = DateFormatter()
+        dateFormatter.dateFormat = "yyyy.MM.dd"
+        return dateFormatter.string(from: Date())
+    }
+}
