--- 90s/90s/Resource/Extension.swift
@@ -1,26 +1,0 @@
-//
-//  Extension.swift
-//  90s
-//
-//  Created by 성다연 on 2021/02/20.
-//
-
-import UIKit
-import Foundation
-
-extension Date {
-    func dateToString() -> String {
-        let dateFormatter = DateFormatter()
-        dateFormatter.dateFormat = "yyyy.MM.dd"
-        return dateFormatter.string(from: Date())
-    }
-}
-
-extension UIColor {
-    class func colorRGBHex(hex:Int, alpha: Float = 1.0) -> UIColor {
-        let r = Float((hex >> 16) & 0xFF)
-        let g = Float((hex >> 8) & 0xFF)
-        let b = Float((hex) & 0xFF)
-        return UIColor(red: CGFloat(r/255.0), green: CGFloat(g/255.0), blue:CGFloat(b/255.0), alpha : CGFloat(alpha))
-    }
-}
