+++ 90s/90s/Resource/NavBarRightBtn.swift
@@ -9,28 +9,28 @@ import Foundation
 import UIKit
 
 enum NavBarRightBtn {
-    case text_cancle
-    case text_edit
-    case img_close
-    case img_check
+    case textCancle
+    case textEdit
+    case imgClose
+    case imgCheck
 }
 
 extension UIButton {
     func setUpNavBarRightBtn(type: NavBarRightBtn){
         switch type {
-        case .text_cancle:
+        case .textCancle:
             setTitleColor( .white, for: .normal)
             titleLabel?.font = .boldSystemFont(ofSize: 16)
             contentEdgeInsets = UIEdgeInsets(top: 0, left: 0, bottom: 0, right: 5)
             setTitle("취소", for: .normal)
-        case .text_edit:
+        case .textEdit:
             setTitleColor( .white, for: .normal)
             titleLabel?.font = .boldSystemFont(ofSize: 16)
             contentEdgeInsets = UIEdgeInsets(top: 0, left: 0, bottom: 0, right: 5)
             setTitle("편집", for: .normal)
-        case .img_close:
+        case .imgClose:
             setImage(UIImage(named: "navigationBar_close"), for: .normal)
-        case .img_check:
+        case .imgCheck:
             setImage(UIImage(named: "navigationBar_check"), for: .normal)
         }
     }
